+++ DEV/ChicagoPS/Shipping/intg_api_transact.sh
@@ -12,6 +12,8 @@
             #Toggle DataProcessedSuccessfully
 ###############################################################################################################################################
 echo " #### Setting Script Variables"
+DEBUG=true
+ENVIRONMENT="QA"
 CLIENT="CPS"
 TYPE="Shipping"
 REGION="us-east-1"
@@ -40,8 +42,12 @@ hayes-datamapper --complete -id $INTEGRATIONID;
 echo " #### $TYPE Data Push process complete!"
 
 echo " #### Sending completion email ";
-# RECIPIENTS="ToAddresses=""support@hayessoft.com"",CcAddresses=""jayala@hayessoft.com,gcollazo@hayessoft.com""";
-RECIPIENTS="ToAddresses=""lsager@hayessoft.com, gcollazo@hayessoft.com"",CcAddresses=""jayala@hayessoft.com""";
+if [ $ENVIRONMENT = "Production" ]; then
+    RECIPIENTS="ToAddresses=""support@hayessoft.com"",CcAddresses=""jayala@hayessoft.com,gcollazo@hayessoft.com,lsager@hayessoft.com""";
+else
+    # RECIPIENTS="ToAddresses=""lsager@hayessoft.com, gcollazo@hayessoft.com"",CcAddresses=""jayala@hayessoft.com""";
+    RECIPIENTS="ToAddresses=""gcollazo@hayessoft.com""";
+fi
 TEXTCONTENT="\nThe $TYPE Integration has completed.\n\nTo access the results go to the Integration Portal and select Instance $INSTANCEID\n\nIf you have any questions please contact support at 1-800-495-5993 or support@hayessoft.com\n\nHayes Software Systems";
 HTMLCONTENT="<br />The $TYPE Integration has completed.<br /><br />To access the results go to the Integration Portal and select Instance $INSTANCEID<br /><br />If you have any questions please contact support at 1-800-495-5993 or support@hayessoft.com<br /><br />Hayes Software Systems";
 MESSAGE="Subject={Data=""$CLIENT $TYPE Integration Status - $CURRENTDATE"",Charset=""ascii""},Body={Text={Data=$TEXTCONTENT,Charset=""utf8""},Html={Data=$HTMLCONTENT,Charset=""utf8""}}";
@@ -49,5 +55,8 @@ MESSAGE="Subject={Data=""$CLIENT $TYPE Integration Status - $CURRENTDATE"",Chars
 aws ses send-email --from "do_not_reply@hayessoft.com" --destination "$RECIPIENTS" --message "$MESSAGE";
 
 echo " #### Terminate instance $INSTANCEID"
-aws ec2 terminate-instances --instance-ids $INSTANCEID
-# aws ec2 stop-instances --instance-ids $INSTANCEID
+if [ ! $DEBUG ]; then
+    aws ec2 terminate-instances --instance-ids $INSTANCEID
+else
+    aws ec2 stop-instances --instance-ids $INSTANCEID
+fi
